requirements.txt
pytest==6.2.4
selenium==4.0.0
webdriver-manager==3.5.2
pytest-env==0.6.2
requests==2.26.0
allure-pytest==2.9.43
jsonschema~=4.4.0
mimesis~=5.3.0